% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfive.R
\name{addfive}
\alias{addfive}
\title{Adds five or 5 to the str/num}
\usage{
addfive(x, delim = " ")
}
\arguments{
\item{x}{is string or numeric value}

\item{delim}{is the deliminator if string, default is " "}
}
\value{
returns a string if x is a string, returns a numeric value if x is numeric
}
\description{
Adds five or 5 to the str/num
}
