% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnum.R
\name{addition}
\alias{addition}
\title{adds number to another number}
\usage{
addition(x, num)
}
\arguments{
\item{x}{is the second number to add}

\item{num}{is the first number to add}
}
\value{
a string
}
\description{
adds number to another number
}
