% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnums.R
\name{addnums}
\alias{addnums}
\title{Adds things together strings or simple addition}
\usage{
addnums(x, num, type = "string", delim = "")
}
\arguments{
\item{x}{is the first string or number}

\item{num}{is the number to add (should be numeric)}

\item{type}{should be string or numeric, determines output}

\item{delim}{if type is string, this separates the x and num}
}
\value{
a string or numeric value
}
\description{
Adds things together strings or simple addition
}
