% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnum.R
\name{addToStr}
\alias{addToStr}
\title{Adds number as string}
\usage{
addToStr(x, num, delim = "")
}
\arguments{
\item{x}{is the string to be added to; must be a character}

\item{num}{the number added to the string (numeric value will be changed to word form)}

\item{delim}{the separator, default is ""}
}
\value{
a string string
}
\description{
Adds number as string
}
