% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regpoly_area.R
\name{regpoly_area}
\alias{regpoly_area}
\title{Calculates area of a regular polygon when given the number and length of the sides}
\usage{
regpoly_area(side_count, length)
}
\arguments{
\item{side_count}{is a numeric value}

\item{length}{is a numeric value}
}
\value{
a numeric value
}
\description{
Calculates area of a regular polygon when given the number and length of the sides
}
