% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfive.R
\name{strfive}
\alias{strfive}
\title{Add the word five to the string}
\usage{
strfive(str, delim = " ")
}
\arguments{
\item{str}{is a vector}

\item{delim}{is deliminator, default is " "}
}
\value{
returns a string
}
\description{
Add the word five to the string
}
